Refactoring to POM test_environment_set_freestyle_project.py
