+++ pages/base_page.py
@@ -72,3 +72,9 @@ class BasePage:
 
     def click_on(self, locator, timeout=5):
         self._wait_for(locator, EC.element_to_be_clickable, timeout).click()
+
+    def scroll_to_element(self, By, Selector):
+        from selenium.webdriver import ActionChains
+        actions = ActionChains(self.driver)
+        actions.move_to_element(self.find_element(By, Selector)).perform()
+        return self
